Added TAP test
